requirements.txt
decorator==4.0.11
django==1.9
djangorestframework==3.5.3
google-api-python-client==1.6.2
